Update IceGrid templates to make the OMERO.tables module configurable
Add a new top-level TABLES_MODULE variable used by the Tables template
to start the new server.
Functionally, this will need to be tested alongside a corresponding
OMERO.py with the logic to replace @omero.tables.module@ by either the
value of the omero.tables.module configuration or the default runTables
value.