Supplement TestTables.testColumnSpaceNames with advanced queries
Variable mapping passed to getWhereList can be used both for the
column names and the values. The current test logic was focusing
on the checking the replacement of column names with spaces.
This commit adds a few additional advanced scenarios where the
integer value used for the condition is also mapped using the supplied
dictionary